0215 f h

+++ SuperMinersServerApplication/MetaData/SystemConfig/GameConfig.cs
@@ -50,7 +50,7 @@ namespace MetaData.SystemConfig
         /// 多少积分等价于1RMB
         /// </summary>
         [DataMember]
-        public decimal Credits_RMB = 1;
+        public int Credits_RMB = 1;
 
         /// <summary>
         /// 每个矿工每小时生产矿石数
